Load initial ingredients from the server
I've handled my error differently from the tutorial.
Instead of showing a second permanent error, I set the ingredients to 0.

Because ingredients is initially null, much of the burger builder needed
to be adapted to be tolerent of nulls. It feels like there should be
another component that can rely on ingredients not being null.